+++ docs/changelog.rst
@@ -3,6 +3,7 @@ scrapelib changelog
 
 1.1.0-dev
 -----
+    * large refactor of billy.site.{browse,www} into billy.web.{admin,public}
     * improved vote validation during scrape
 
 1.0.0